% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{rfuns_sum}
\alias{rfuns_sum}
\alias{rfuns_min}
\alias{rfuns_max}
\title{aggregate functions}
\usage{
rfuns_sum(x, ...)

rfuns_min(x, ...)

rfuns_max(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{passed down, e.g. na.rm}
}
\description{
aggregate functions
}
\examples{
rfuns_sum(1:10, na.rm = FALSE)
rfuns_min(1:10, na.rm = FALSE)
rfuns_max(1:10, na.rm = FALSE)

}
