% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_con.R
\name{con}
\alias{con}
\title{Open a connection to duckdb with the rfuns extension loaded}
\usage{
con()
}
\value{
a duckdb connection
}
\description{
Open a connection to duckdb with the rfuns extension loaded
}
