% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuns.R
\name{rfuns}
\alias{rfuns}
\title{Call an rfuns function}
\usage{
rfuns(
  op = c("project", "aggregate"),
  fun,
  data,
  ...,
  error_call = caller_env()
)
}
\arguments{
\item{op}{project or aggregate}

\item{fun}{function name, without the r_base:: prefix}

\item{data}{data frame}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{dots}{constants}
}
\description{
Call an rfuns function
}
\examples{
rfuns("aggregate", "sum", data.frame(x = 1:10))

}
