% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{rfuns_sum}
\alias{rfuns_sum}
\title{sum}
\usage{
rfuns_sum(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector}

\item{na.rm}{should the missing values be removed}
}
\description{
sum
}
\examples{
rfuns_sum(1:10)

}
